+++ include/sora/sora_signaling.h
@@ -158,6 +158,8 @@ struct SoraSignalingConfig {
   bool disable_signaling_url_randomization = false;
 
   std::optional<http_header_value> user_agent;
+
+  std::optional<webrtc::DegradationPreference> degradation_preference;
 };
 
 class SoraSignaling : public std::enable_shared_from_this<SoraSignaling>,
@@ -226,6 +228,9 @@ class SoraSignaling : public std::enable_shared_from_this<SoraSignaling>,
   void SetEncodingParameters(
       std::string mid,
       std::vector<webrtc::RtpEncodingParameters> encodings);
+  void SetDegradationPreference(
+      std::string mid,
+      webrtc::DegradationPreference degradation_preference);
   void ResetEncodingParameters();
 
   void WsWriteSignaling(std::string text, Websocket::write_callback_t on_write);