+++ test/hello.h
@@ -19,6 +19,7 @@ struct HelloSoraConfig {
   std::string client_id;
   std::vector<sora::SoraSignalingConfig::DataChannel> data_channels;
   std::vector<sora::SoraSignalingConfig::ForwardingFilter> forwarding_filters;
+  std::optional<webrtc::DegradationPreference> degradation_preference;
 };
 
 class HelloSora : public std::enable_shared_from_this<HelloSora>,