+++ CHANGES.md
@@ -71,6 +71,10 @@
   - @melpon
 - [ADD] sumomo にビデオデバイスとオーディオデバイスを設定するオプションを追加
   - @melpon
+- [ADD] SoraSignalingConfig に degradation_preference を追加 
+  - @melpon
+- [ADD] sumomo と test/hello に degradation_preference を設定するオプションを追加
+  - @melpon
 - [FIX] HTTP Proxy 利用時の Websocket 初期化で insecure_ メンバ変数が初期化されていなかったのを修正
   - @melpon
 - [FIX] SoraSignalingConfig の client_cert と client_key に渡す必要がある値を、ファイルパスからファイルの内容に修正